+++ config/config.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
-	"strings"
 
 	"github.com/spf13/viper"
 )
@@ -71,26 +70,4 @@ func InferSiteURL(apiURL string) string {
 	}
 	re := regexp.MustCompile("^(https?://[^/]*).*")
 	return re.ReplaceAllString(apiURL, "$1")
-}
-
-func Resolve(path, home string) string {
-	if path == "" {
-		return ""
-	}
-	if strings.HasPrefix(path, "~/") {
-		path = strings.Replace(path, "~/", "", 1)
-		return filepath.Join(home, path)
-	}
-	if filepath.IsAbs(path) {
-		return filepath.Clean(path)
-	}
-	// if using "/dir" on Windows
-	if strings.HasPrefix(path, "/") {
-		return filepath.Join(home, filepath.Clean(path))
-	}
-	cwd, err := os.Getwd()
-	if err != nil {
-		return path
-	}
-	return filepath.Join(cwd, path)
 }
